templates/django/server/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
pyton-dehcouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0